Changes to make roi_table work for all datasets
Changed the way the ROI table is made to use 'fetch_primary_rois' to find the ROIs to include.
It also uses `fetch_adjacencies` not `fetch_neurons` to find the number of connections between all cells of the type.
A function `has_soma_side_column` was added to the dataset adaptors to return whether or not the somaSide column exists in the dataset. This changes which fields are used in the NeuronCriteria for fetch_adjacencies when making the ROI table.
The html template was updated to match the output of the ROI_table function. 'Pre' and 'post' were changed for 'inputs' and 'outputs' to avoid confusion with synapse counts since these numbers refer to connections.